app/util/cprbroker/models/Relationship.java
/* ***** BEGIN LICENSE BLOCK *****
 * Version: MPL 2.0/GPL 2.0/LGPL 2.1
 *
 * The contents of this file are subject to the Mozilla Public License
 * Version 2.0 (the "License"); you may not use this file except in
 * compliance with the License. You may obtain a copy of the License at
 * http://www.mozilla.org/MPL/
 *
 * Software distributed under the License is distributed on an "AS IS"basis,
 * WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
 * for the specific language governing rights and limitations under the
 * License.
 *
 * Contributor(s):
 * Søren Kirkegård
 *
 * The code is currently governed by OS2 - Offentligt digitaliserings-
 * fællesskab / http://www.os2web.dk .
 *
 * Alternatively, the contents of this file may be used under the terms of
 * either the GNU General Public License Version 2 or later (the "GPL"), or
 * the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
 * in which case the provisions of the GPL or the LGPL are applicable instead
 * of those above. If you wish to allow use of your version of this file only
 * under the terms of either the GPL or the LGPL, and not to allow others to
 * use your version of this file under the terms of the MPL, indicate your
 * decision by deleting the provisions above and replace them with the notice
 * and other provisions required by the GPL or the LGPL. If you do not delete
 * the provisions above, a recipient may use your version of this file under
 * the terms of any one of the MPL, the GPL or the LGPL.
 *
 * ***** END LICENSE BLOCK ***** */

package util.cprbroker.models;

import util.cprbroker.ERelationshipType;
import util.cprbroker.IRelationship;
import util.cprbroker.IVirkning;

public class Relationship implements IRelationship {

	private final String comment;
	private final String referenceUrn;
	private final String referenceUuid;
	private final IVirkning effect;
	private final ERelationshipType type;
	
	public static class Builder{
		
		private String comment;
		private String referenceUrn;
		private String referenceUuid;
		private IVirkning effect;
		private ERelationshipType type;
		
		public IRelationship build() { return new Relationship(this); }
		
		public Builder comment(String newComment) { comment = newComment; return this; }
		public Builder referenceUrn(String newUrn) { referenceUrn = newUrn; return this; }
		public Builder referenceUuid(String newUuid) { referenceUuid = newUuid; return this; }
		public Builder effect(IVirkning newEffect) { effect = newEffect; return this; }
		public Builder type(ERelationshipType newType) { type = newType; return this; }
		
	}
	
	private Relationship(Builder builder) {
		comment = builder.comment;
		referenceUrn = builder.referenceUrn;
		referenceUuid = builder.referenceUuid;
		effect = builder.effect;
		type = builder.type;
	}
	
	@Override
	public String comment() { return comment; }

	@Override
	public String referenceUrn() { return referenceUrn; }

	@Override
	public String referenceUuid() {	return referenceUuid; }

	@Override
	public IVirkning effect() { return effect; }
	
	@Override
	public ERelationshipType relationshipType() { return type;}

}
